% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{create_metrics_req}
\alias{create_metrics_req}
\title{Create metrics request}
\usage{
create_metrics_req(...)
}
\arguments{
\item{...}{further arguments to httr's GET.}
}
\value{
A modified HTTP request.
}
\description{
Create metrics request
}
\keyword{internal}
