% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{create_req}
\alias{create_req}
\title{Create request}
\usage{
create_req(...)
}
\arguments{
\item{...}{further arguments to httr's GET.}
}
\value{
A modified HTTP request.
}
\description{
Create request
}
\keyword{internal}
