% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error_parameter_is_null}
\alias{error_parameter_is_null}
\title{Throw an error if parameter is null}
\usage{
error_parameter_is_null(parameters = NULL)
}
\arguments{
\item{parameters}{List of parameters.}
}
\value{
Error message if parameters missing.
}
\description{
Throw an error if parameter is null
}
