% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Create metrics request}
\usage{
get_metrics(...)
}
\arguments{
\item{...}{further arguments to httr's GET.}
}
\value{
A modified HTTP request.
}
\description{
Create metrics request
}
\keyword{internal}
