% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageviews.R
\name{get_most_viewed_per_country}
\alias{get_most_viewed_per_country}
\title{Get most viewed articles per country}
\usage{
get_most_viewed_per_country(
  country = "CA",
  access = "all-access",
  year = "2022",
  month = "01",
  day = "01"
)
}
\arguments{
\item{country}{description}

\item{access}{all-access, desktop, mobile-app, mobile-web}
}
\value{
A modified HTTP request.
}
\description{
Get most viewed articles per country
}
\keyword{internal}
