% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikirest-package.R
\docType{package}
\name{wikirest-package}
\alias{wikirest}
\alias{wikirest-package}
\title{wikirest: Get data from the Wikimedia REST API}
\description{
Get data from the Wikimedia REST API.
}
\author{
\strong{Maintainer}: Judith Bourque \email{info@clessn.ca}

}
\keyword{internal}
