% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{add_user_agent}
\alias{add_user_agent}
\title{Add user agent to request}
\usage{
add_user_agent(req)
}
\arguments{
\item{req}{A HTTP request.}
}
\value{
A modified HTTP request.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
