% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{create_metrics_req}
\alias{create_metrics_req}
\title{Create metrics request}
\usage{
create_metrics_req()
}
\value{
A modified HTTP request.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
