% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageviews.R
\name{create_pageviews_req}
\alias{create_pageviews_req}
\title{Create pageviews request}
\usage{
create_pageviews_req()
}
\value{
A modified HTTP request.
}
\description{
Create pageviews request
}
\keyword{internal}
