% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{create_req}
\alias{create_req}
\title{Create request}
\usage{
create_req()
}
\value{
A modified HTTP request.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
