% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error_if_parameter_null}
\alias{error_if_parameter_null}
\title{Throw an error if parameter is null}
\usage{
error_if_parameter_null(parameters = NULL)
}
\arguments{
\item{parameters}{List or vector of parameters.}
}
\value{
Error message if parameters missing.
}
\description{
Throw an error if parameter is null
}
\keyword{internal}
