% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageviews.R
\name{get_most_viewed_per_country}
\alias{get_most_viewed_per_country}
\title{Get most viewed articles per country}
\usage{
get_most_viewed_per_country(
  country = NULL,
  access = NULL,
  date = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  tidy = TRUE
)
}
\arguments{
\item{country}{String. ISO 3166-1 alpha-2 code of a country}

\item{access}{String. One of the following: all-access, desktop, mobile-app,
mobile-web.}

\item{date}{String. Date in YYYY-MM-DD format.}

\item{year}{String. Year in YYYY format.}

\item{month}{String. Month in MM format.}

\item{day}{String. Day in DD format.}

\item{tidy}{Logical. TRUE to return a tibble. FALSE to return an httr2_response.}
}
\value{
If tidy = TRUE, a tibble, if false, an httr2_response.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
get_most_viewed_per_country(country = "CA", access = "all-access",
  year = "2022", month = "01", day = "01", tidy = TRUE)
}
